Prepare for ClojureScript development
We will hand-port the CLJS version of the main tools.cli namespace, but
use lein-cljsbuild, lein-dalap, and clojurescript.test for testing both
CLJ and CLJS builds.

lein-dalap is a preprocessor like cljx, but has the advantage that
CLJS-only forms are hidden within #_(:cljs ) comment forms so that the
file may remain as valid Clojure.

This is only for local testing of the hand-crafted CLJS file, so I don't
believe this runs afoul of any of the contrib library guidelines.