Handle nil long-opt in spec when checking nil-flag
The addition of the `nil-flag?` check did not handle `nil` values for
`:long-opt` in the spec.  This causes a `NullPointerException`.  This
commit checks that there is a long-opt in the spec before testing it for
the `"--no"` string.

Addtionally, modifies a unit test, such that one of the options has a
`nil` long opt, to prevent future regressions.

Signed-off-by: Peter Schwarz <peter.schwarz@bicycle.io>
Signed-off-by: Sean Corfield <sean@corfield.org>